Move logstash filter code from dao to facade
